fix: unstage reserved scratch before prepare dirt checks
Unstage newly-added reserved reviewer scratch in PrepareWorktree so prestaged
payloads under .gitignore negation no longer stick retry in manual_intervention.
Fail closed when git status capture is truncated so filtered reserved-only
prefixes cannot report clean while ordinary dirt was omitted. Split limit-edge
contracts into a separate test file under the 300-line threshold.

Generated-By: looper 0.11.2 (runner=fixer, agent=grok-build)
