+++ internal/infra/git/gateway_reserved_scratch_test.go
@@ -2,7 +2,6 @@ package git
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -38,7 +37,7 @@ func TestIsReservedReviewerScratchPath(t *testing.T) {
 
 // Dual-classifier real-Git contracts for reserved /.looper-review-*.json.
 // Basename edges live in the unit table; cases below cover prepare/exclude,
-// negation, prestaged, rename, tracked M, and literal pathspec unstage.
+// negation, prestaged prepare+commit, rename, tracked M, and literal pathspecs.
 func TestGatewayReservedReviewerScratchContract(t *testing.T) {
 	type caseSpec struct {
 		name, branch    string
@@ -84,7 +83,6 @@ func TestGatewayReservedReviewerScratchContract(t *testing.T) {
 				writeFile(t, filepath.Join(wt, ".looper-review-.json"), "{}\n")
 				writeFile(t, filepath.Join(wt, ".looper-review-1049.json"), "{}\n")
 			},
-			// Nested lookalike is commit-only dirt so Prepare can stay clean under negation.
 			beforeCommit: func(t *testing.T, wt string) {
 				mustMkdirAll(t, filepath.Join(wt, "nested"))
 				writeFile(t, filepath.Join(wt, "nested", ".looper-review-1.json"), "{}\n")
@@ -94,12 +92,14 @@ func TestGatewayReservedReviewerScratchContract(t *testing.T) {
 			keepOnDisk: []string{".looper-review-1049.json", ".looper-review-.json"},
 		},
 		{
-			name: "commit_unstages_prestaged", branch: "feature/review-prestaged", gitignoreNegate: true,
+			// Prestaged under negation must unstage in Prepare (retry lifecycle),
+			// then Commit must still keep reserved scratch off HEAD.
+			name: "prepare_and_commit_unstages_prestaged", branch: "feature/review-prestaged", gitignoreNegate: true,
 			setup: func(t *testing.T, wt string) {
 				writeFile(t, filepath.Join(wt, ".looper-review-1050.json"), "{}\n")
 				runGit(t, wt, "add", "-A", "--", ".looper-review-1050.json")
 			},
-			include: []string{"app.go"}, exclude: []string{".looper-review-1050.json"},
+			wantClean: &yes, include: []string{"app.go"}, exclude: []string{".looper-review-1050.json"},
 			keepOnDisk: []string{".looper-review-1050.json"},
 		},
 		{
@@ -227,94 +227,6 @@ func TestGatewayReservedReviewerScratchContract(t *testing.T) {
 				}
 			}
 		})
-	}
-}
-
-// Real-Git: a large ordinary staged-addition listing (>256 KiB of non-scratch
-// paths) must still Commit successfully. Unstage queries only the reserved root
-// namespace, so shell capture bounds cannot reject unrelated worker/planner/fixer
-// publishes while still keeping reserved scratch off the commit.
-func TestGatewayCommitSucceedsWithLargeNonScratchAdditionListing(t *testing.T) {
-	ctx := context.Background()
-	fixture := newFixture(t)
-	branch := "feature/review-large-listing"
-	fixture.createRemoteRepo(t, branch)
-	gateway := fixture.gateway()
-
-	// Higher-precedence negation so Commit's git add -A stages reserved scratch;
-	// info/exclude alone would hide it and skip the unstage path under test.
-	runGit(t, fixture.repoPath, "checkout", branch)
-	writeFile(t, filepath.Join(fixture.repoPath, ".gitignore"), "!/.looper-review-*.json\n")
-	runGit(t, fixture.repoPath, "add", ".gitignore")
-	runGit(t, fixture.repoPath, "commit", "-m", "seed negation for large listing")
-	runGit(t, fixture.repoPath, "push", "origin", branch)
-	runGit(t, fixture.repoPath, "checkout", "main")
-
-	worktree, err := gateway.CreateWorktree(ctx, CreateWorktreeInput{
-		ProjectID: fixture.projectID, RepoPath: fixture.repoPath, WorktreeRoot: fixture.worktreeRoot,
-		Branch: branch, BaseBranch: "main", PRNumber: 1048,
-	})
-	if err != nil {
-		t.Fatalf("CreateWorktree() error = %v", err)
-	}
-	wt := worktree.WorktreePath
-	writeFile(t, filepath.Join(wt, "app.go"), "package main\n")
-	writeFile(t, filepath.Join(wt, ".looper-review-large.json"), "{}\n")
-	// Long basenames: ~1500 additions exceed defaultMaxOutputBytes of -z --name-only
-	// for an unscoped listing; pathspec-scoped scratch query stays tiny.
-	bulkDir := filepath.Join(wt, "bulk")
-	mustMkdirAll(t, bulkDir)
-	pad := strings.Repeat("x", 180)
-	for i := 0; i < 1500; i++ {
-		writeFile(t, filepath.Join(bulkDir, fmt.Sprintf("f%04d_%s.txt", i, pad)), "1\n")
-	}
-	runGit(t, wt, "add", "-A")
-	fullListing, err := runGitCommand(wt, "diff", "--cached", "--name-only", "--diff-filter=A", "--no-renames", "-z")
-	if err != nil {
-		t.Fatalf("diff --cached full listing error = %v", err)
-	}
-	if len(fullListing) <= 256*1024 {
-		t.Fatalf("fixture full listing is %d bytes; need >256KiB to prove unscoped capture would truncate", len(fullListing))
-	}
-	if !strings.Contains(fullListing, ".looper-review-large.json") {
-		t.Fatalf("fixture full listing missing staged reserved scratch under negation")
-	}
-	scopedListing, err := runGitCommand(wt, "diff", "--cached", "--name-only", "--diff-filter=A", "--no-renames", "-z",
-		"--", ":(glob).looper-review-*.json")
-	if err != nil {
-		t.Fatalf("diff --cached scoped listing error = %v", err)
-	}
-	if len(scopedListing) == 0 || !strings.Contains(scopedListing, ".looper-review-large.json") {
-		t.Fatalf("scoped listing missing reserved scratch; got %q", scopedListing)
-	}
-	if len(scopedListing) > 256*1024 {
-		t.Fatalf("scoped listing unexpectedly huge (%d bytes)", len(scopedListing))
-	}
-
-	if _, err := gateway.Commit(ctx, CommitInput{
-		RepoPath: fixture.repoPath, WorktreeRoot: fixture.worktreeRoot,
-		WorktreePath: wt, Message: "large non-scratch commit must publish",
-	}); err != nil {
-		t.Fatalf("Commit() error = %v, want success for large non-scratch listing", err)
-	}
-	if _, statErr := os.Stat(filepath.Join(wt, ".looper-review-large.json")); statErr != nil {
-		t.Fatalf("expected reserved scratch preserved on disk: %v", statErr)
-	}
-	committed := runGit(t, wt, "diff-tree", "--no-commit-id", "--name-only", "-r", "-z", "HEAD")
-	if strings.Contains(committed, ".looper-review-large.json") {
-		t.Fatalf("HEAD unexpectedly contains reserved scratch; files = %q", committed)
-	}
-	if !strings.Contains(committed, "app.go") {
-		t.Fatalf("HEAD missing app.go; files = %q", committed)
-	}
-	// Sample one bulk path; full 1500-name check is unnecessary.
-	sampleBulk := fmt.Sprintf("bulk/f0000_%s.txt", pad)
-	if !strings.Contains(committed, sampleBulk) {
-		prefix := committed
-		if len(prefix) > 200 {
-			prefix = prefix[:200]
-		}
-		t.Fatalf("HEAD missing bulk sample %q; files prefix = %q", sampleBulk, prefix)
 	}
 }
